refactor multi-cloze
